@@ -11,24 +11,27 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/busboy": "^0.2.3",
     "@types/express": "^4.17.11",
     "@types/sequelize": "^4.28.9",
+    "@types/uuid": "^8.3.0",
     "jest": "^26.6.3",
     "ts-jest": "^26.5.1",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.5"
+    "typescript": "^4.2.2"
   },
   "dependencies": {
     "axios": "^0.21.1",
+    "busboy": "^0.3.1",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "form-data": "^4.0.0",
     "inversify": "^5.0.5",
     "pg": "^8.5.1",
     "pg-hstore": "^2.3.3",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^6.5.0",
     "sequelize-typescript": "^2.1.0",
-    "sqlite3": "^5.0.2",
-    "tedious": "^11.0.4",
     "uuid": "^8.3.2"
   }
 }
