@@ -1,12 +1,14 @@
 import { inject, injectable } from "inversify";
-import TYPES from "../ioc-container/types";
-import { IPost } from "../models/post.model";
 import { IPostRepository } from "../repositories/post.repository";
+import { IImageUploader } from "./image.uploader";
+import { IPost } from "../models/post.model";
+import TYPES from "../ioc-container/types";
+import uuid from 'uuid';
 
 
 export interface IPostService {
-    addPost(post: IPost): Promise<IPost>;
-    removePostById(postId: string): Promise<boolean>;
+    addPost(post: IPost, uploadFile: object, accessToken: string): Promise<IPost>;
+    removePostById(postId: string, accessToken: string): Promise<boolean>;
     getFilteredPosts(userFilter: string[], tagFilter: string[], publishers: string[], distance: number, from: Date, to: Date): Promise<IPost[]>;
     getPostDataById(postId: string): Promise<IPost>;
     getAllPostsByUserId(userId: string): Promise<IPost[]>;
@@ -15,7 +17,8 @@ export interface IPostService {
 
 @injectable()
 export class PostService implements IPostService {
-    constructor(@inject(TYPES.IPostRepository) private repository: IPostRepository) {
+    constructor(@inject(TYPES.IPostRepository) private repository: IPostRepository,
+        @inject(TYPES.IImageUploader) private uploader: IImageUploader) {
         this.addPost = this.addPost.bind(this);
         this.removePostById = this.removePostById.bind(this);
         this.getFilteredPosts = this.getFilteredPosts.bind(this);
@@ -24,20 +27,26 @@ export class PostService implements IPostService {
         this.updatePost = this.updatePost.bind(this);
     }
 
-    addPost(post: IPost): Promise<IPost> {
+    async addPost(post: IPost, uploadFile: any, accessToken: string): Promise<IPost> {
         if (!post) {
             throw new ReferenceError('No post proivded!');
         }
 
+        post.id = uuid.v4();
+        post.imageUrl = await this.uploader.uploadImage(post.publisherId, post.id, uploadFile, accessToken);
+
         return this.repository.addPost(post);
     }
 
-    removePostById(postId: string): Promise<boolean> {
+    async removePostById(postId: string, accessToken: string): Promise<boolean> {
         if (!postId) {
             throw new ReferenceError('No post id proivded!');
         }
 
-        return this.repository.removePost(postId);
+        const post = await this.repository.removePost(postId);
+        const success = await this.uploader.deleteImage(post.imageUrl, accessToken);
+
+        return post != undefined;
     }
 
     getFilteredPosts(userFilter: string[], tagFilter: string[], publishers: string[], distance: number, from: Date, to: Date): Promise<IPost[]> {
