@@ -1,17 +1,10 @@
 export default {
-    PORT: 5000,
     jwtSettings: {
         accessToken: {
             expiration: '15m',
         },
         refreshToken: {
             expiration: '1M',
-        },
-        secret: 'QfTjWnZr4u7x!z%C*F-JaNdRgUkXp2s5v8y/B?E(G+KbPeShVmYq3t6w9z$C&F)J@McQfTjWnZr4u7x!A%D*G-KaPdRgUkXp2s5v8y/B?E(H+MbQeThVmYq3t6w9z$C&F)J@NcRfUjXnZr4u7x!A%D*G-KaPdSgVkYp3s5v8y/B?E(H+MbQeThWmZq4t7w9z$C&F)J@NcRfUjXn2r5u8x/A%D*G-KaPdSgVkYp3s6v9y$B&E(H+MbQeThWmZq4t7'
-    },
-    googleSettings: {
-        clientId: '332337525586-rpk6kqpc36ja5st3dsc40g8593e5kjkj.apps.googleusercontent.com',
-        clientSecret: 'FGU48bljpZBBoGkTsmcQJ0s8',
-        callbackUrl: 'http://localhost:5000/google/callback' 
+        }
     }
 };
