@@ -6,8 +6,16 @@ const TYPES = {
 
     IPostService: Symbol('IPostService'),
     ICommentService: Symbol('ICommentService'),
+    IUserTagService: Symbol('IUserTagService'),
+    ITagService: Symbol('ITagService'),
+    IImageUploader: Symbol('IImageUploader'),
     
-    Sequelize: Symbol('Sequelize')
+    Sequelize: Symbol('Sequelize'),
+
+    PostController: Symbol('PostController'),
+    CommentController: Symbol('CommentController'),
+    UserTagController: Symbol('UserTagController'),
+    TagController: Symbol('TagController'),
 }
 
 export default TYPES;
